chore: Update Docker workflow to set working directory for build step
Update the Docker workflow to set the working directory for the build step to "./src". This ensures that the correct directory is used when building the abpreact nextjs image. Previously, it was set to "./apps/abp", which was incorrect. This change aligns the workflow with the actual file structure and improves the accuracy of the build process.